requirements.txt
flake8==6.0.0
pre-commit==3.3.3
django==4.2.3
environs==9.5.0
django-phonenumber-field==7.1.0
twilio==8.5.0
